+++ pizzeria/app/Http/Controllers/api/OrderPizzaController.php
@@ -24,6 +24,12 @@ class OrderPizzaController extends Controller
 
         return $orderPizzas;
     }
+
+    private $validate = Validator::make($request->all(), [
+        'order_id' => 'required|exists:orders,id',
+        'pizza_size_id' => 'required|exists:pizza_size,id',
+        'quantity' => 'required|integer|min:1'
+    ]);
     /**
      * Display a listing of the resource.
      */
@@ -37,6 +43,13 @@ class OrderPizzaController extends Controller
      */
     public function store(Request $request)
     {
+        if ($this->validate->fails()) {
+            return response()->json([
+                'msg' => 'Se produjo un error en la validación de la información.',
+                'statusCode' => 400
+            ]);
+        }
+
         $orderPizza = new order_pizza();
 
         $orderPizza->order_id = $request->order_id;
@@ -54,6 +67,10 @@ class OrderPizzaController extends Controller
     {
         $orderPizza = order_pizza::find($id);
 
+        if (is_null($orderPizza)) {
+            return abort(404);
+        }
+
         return json_encode($orderPizza);
     }
 
@@ -62,7 +79,18 @@ class OrderPizzaController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if ($this->validate->fails()) {
+            return response()->json([
+                'msg' => 'Se produjo un error en la validación de la información.',
+                'statusCode' => 400
+            ]);
+        }
+
         $orderPizza = order_pizza::find($id);
+
+        if (is_null($orderPizza)) {
+            return abort(404);
+        }
 
         $orderPizza->order_id = $request->order_id;
         $orderPizza->pizza_size_id = $request->pizza_size_id;
@@ -78,6 +106,11 @@ class OrderPizzaController extends Controller
     public function destroy(string $id)
     {
         $orderPizza = order_pizza::find($id);
+
+        if (is_null($orderPizza)) {
+            return abort(404);
+        }
+
         $orderPizza->delete();
 
         return json_encode($this->getOrdersPizzas());
